break up the main component
