+++ app/packages/core/src/components/Actions/ActionsRow.tsx
@@ -35,8 +35,8 @@ import { GroupMediaVisibilityContainer } from "./GroupMediaVisibilityContainer";
 import OptionsActions from "./Options";
 import Patcher, { patchesFields } from "./Patcher";
 import Selector from "./Selected";
-import Similar, { isImageSimilaritySearch } from "./Similar";
 import Tagger from "./Tagger";
+import Similar, { isImageSimilaritySearch } from "./similar/Similar";
 
 export const shouldToggleBookMarkIconOnSelector = selector<boolean>({
   key: "shouldToggleBookMarkIconOn",
